backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
requests==2.31.0
python-dotenv==1.0.0
django-extensions==3.2.3
factory-boy==3.3.0
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0
django-filter==23.5
drf-spectacular==0.26.5
django-redis==5.4.0
